+++ src/main/java/com/rebook/review/service/dto/ReviewDto.java
@@ -1,4 +1,4 @@
-package com.rebook.review.dto;
+package com.rebook.review.service.dto;
 
 import com.rebook.book.domain.entity.BookEntity;
 import com.rebook.review.domain.ReviewEntity;
@@ -12,15 +12,15 @@ import java.time.ZonedDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
-public class ReviewDTO {
+public class ReviewDto {
     private BookEntity book;
     private Long id;
     private String content;
     private BigDecimal starRate;
     private ZonedDateTime createdAt;
 
-    public static ReviewDTO fromEntity(ReviewEntity reviewEntity) {
-        return new ReviewDTO(reviewEntity.getBook(),
+    public static ReviewDto fromEntity(ReviewEntity reviewEntity) {
+        return new ReviewDto(reviewEntity.getBook(),
                 reviewEntity.getId(),
                 reviewEntity.getContent(),
                 reviewEntity.getStarRate(),
